requirements.txt
altgraph==0.17.4
keyboard==0.13.5
MouseInfo==0.1.3
packaging==23.2
pefile==2023.2.7
Pillow==10.1.0
PyAutoGUI==0.9.54
PyGetWindow==0.0.9
pyinstaller==6.3.0
pyinstaller-hooks-contrib==2023.11
PyMsgBox==1.0.9
pyperclip==1.8.2
PyRect==0.2.0
PyScreeze==0.1.30
pyserial==3.5
pytweening==1.0.7
pywin32-ctypes==0.2.2
